Adding support for Move on Sui to semgrep (semgrep/semgrep-proprietary#10589) (semgrep/semgrep-proprietary#2425)
This adds initial support for the Sui version of Move to Semgrep.

While this is an initial commit, a large portion of the language works.
Metavariables work, etc.

The big thing missing is support for annotations (eg #test). This
support will be added after it is added to upstream move tree-sitter.

Note: This PR depends on:
https://github.com/semgrep/ocaml-tree-sitter-semgrep/pull/515 to merge
(and a new [semgrep-move-on-sui] to be pushed).
And https://github.com/semgrep/semgrep-interfaces/pull/296 which adds
move-on-sui to the languages.

sync of https://github.com/semgrep/semgrep/pull/10589

synced from Pro fb54fceb72e1ff76440496c15fb70bb01ec2b687