chore: remove default argument for Pos.make (semgrep/semgrep-proprietary#1881)
This commit starts with `Pos.make` removes `~file` from being optional
and propagates those changes through the entire development. In
particular this removes the idea of having fake positions and fake
locations with the hope of decreasing the number of bugs and obviating
the places where we should clean up code. Not sure if this will be
accepted, but I was hoping to start a conversation.

synced from Pro 03cb11be0179db3ed39e7f601304d1fbafb6b7f4