chore: update nix flake every week (semgrep/semgrep-proprietary#2696)
Let's update the nix flakes automatically, every weekend we'll make a PR
that will update the nix flakes, and tag me. This will ensure that if
people add new dependencies, or update them to the newest version, we
won't break the nix workflow.

Note: I updated the nix flakes in this PR, and had to bump cohttp and
fix a breaking api change, but this was small and the change is covered
by tests already.

## Test plan
CI Passes
and also for one of the commits, I had the workflow run on the PR. I
confirmed the PR created
https://github.com/semgrep/semgrep-proprietary/pull/2698 was what was
expected

synced from Pro 63daf8af1a9100bc899403c7a91c7a6ea08efd29