Fix target listing for symbolic links to folders (semgrep/semgrep-proprietary#2784)
The test "scanning root as a symlink to a folder" now passes.

This comes with some simplification in the implementation (listing files
with git is now done by cding into the scanning root after checking that
it's a dir).

synced from Pro 8bcf47532c239255303e6014f99d9423df480b4a