refactor: Add Hook module (semgrep/semgrep-proprietary#2901)
We use hooks in a variety of places to adjust the behavior of some
module without explicitly passing information down. This has proved to
be expedient, in particular, for providing integration points for the
Pro Engine.

Unfortunately, this pattern also makes it very difficult to reason about
the behavior of the program. Some of this difficulty is not fundamental
to the nature of hooks, but instead is due to their representation as
`ref`s. We often set them without later un-setting them, or set them
deep in the call stack with the expectation that the hook will be used
later on. This makes our code difficult to understand and has also led
to problems where hook state is incorrectly left over from a previous
computation.

In the long term, I would prefer to refactor to avoid hooks altogether
where feasible. In the short term, I would at least like to make them
easier to reason about and reduce the opportunity for them to include
stale state.

To these ends, I'm introducing a simple wrapper around `ref`s that only
allows a hook to be set using the `with_hook_set` function. This
function will set the hook to the given value, run the provided
function, and then set the hook back to its previous value. Enforcing
this pattern will make the scope of each updated hook clearer and will
ensure that they are cleaned up when they are no longer needed.

As an example, I have updated
`Match_tainting_mode.hook_setup_hook_function_taint_signature` to use
this new module. I plan to follow up with more hook refactorings.

Test plan: Automated tests

synced from Pro 8baf75383384fb7f78c801990fa67e295e17fae3